Split sync delta merge shapes
