+++ service-worker.js
@@ -156,6 +156,10 @@ const APP_SHELL = [
   '/js/sync-delta-snapshot.js',
   '/js/sync-delta-merge.js',
   '/js/sync-delta-merge-shapes.js',
+  '/js/sync-delta-row-codec.js',
+  '/js/sync-delta-array-merge.js',
+  '/js/sync-delta-map-merge.js',
+  '/js/sync-delta-scalar-merge.js',
   '/js/sync-delta-registry.js',
   '/js/sync-delta-observability.js',
   '/js/sync-tombstones.js',
