+++ tests/test-sync.js
@@ -58,6 +58,10 @@ await import('../js/settings.js');
   const syncDeltaSnapshotSrc = await fetchWithRetry('js/sync-delta-snapshot.js');
   const syncDeltaMergeSrc = await fetchWithRetry('js/sync-delta-merge.js');
   const syncDeltaMergeShapesSrc = await fetchWithRetry('js/sync-delta-merge-shapes.js');
+  const syncDeltaRowCodecSrc = await fetchWithRetry('js/sync-delta-row-codec.js');
+  const syncDeltaArrayMergeSrc = await fetchWithRetry('js/sync-delta-array-merge.js');
+  const syncDeltaMapMergeSrc = await fetchWithRetry('js/sync-delta-map-merge.js');
+  const syncDeltaScalarMergeSrc = await fetchWithRetry('js/sync-delta-scalar-merge.js');
   const syncDeltaRegistrySrc = await fetchWithRetry('js/sync-delta-registry.js');
   const syncDeltaObservabilitySrc = await fetchWithRetry('js/sync-delta-observability.js');
   const syncTombstonesSrc = await fetchWithRetry('js/sync-tombstones.js');
@@ -100,7 +104,8 @@ await import('../js/settings.js');
   const themeExtraSrc = await fetchWithRetry('themes-extra.css');
   const serviceWorkerSrc = await fetchWithRetry('service-worker.js');
   const syncDeltaPlannerSearchSrc = `${syncDeltaPlannersSrc}\n${syncDeltaPlannerContextSrc}\n${syncDeltaArrayPlannerSrc}\n${syncDeltaMapPlannerSrc}\n${syncDeltaScalarPlannerSrc}`;
-  const deltaSearchSrc = `${syncSrc}\n${syncPushSrc}\n${syncPushDeltasSrc}\n${syncReconcileSrc}\n${syncPullSrc}\n${syncPullMergeSrc}\n${syncPullMaintenanceSrc}\n${syncPullActiveRefreshSrc}\n${syncPullRebroadcastSrc}\n${syncCutoverSrc}\n${syncDeltaSrc}\n${syncDeltaPlannerSearchSrc}\n${syncDeltaSnapshotSrc}\n${syncDeltaMergeSrc}\n${syncDeltaMergeShapesSrc}\n${syncDeltaRegistrySrc}\n${syncDeltaObservabilitySrc}\n${syncDiagnosticsSrc}\n${syncDiagnoseActionsSrc}\n${syncDiagnoseActionsContextSrc}\n${syncDiagnoseRelayActionsSrc}\n${syncDiagnoseIdentityActionsSrc}\n${syncDiagnoseCutoverActionsSrc}\n${syncDiagnoseUiSrc}\n${syncDiagnoseRenderSrc}\n${syncWindowBindingsSrc}`;
+  const syncDeltaMergeSearchSrc = `${syncDeltaMergeShapesSrc}\n${syncDeltaRowCodecSrc}\n${syncDeltaArrayMergeSrc}\n${syncDeltaMapMergeSrc}\n${syncDeltaScalarMergeSrc}`;
+  const deltaSearchSrc = `${syncSrc}\n${syncPushSrc}\n${syncPushDeltasSrc}\n${syncReconcileSrc}\n${syncPullSrc}\n${syncPullMergeSrc}\n${syncPullMaintenanceSrc}\n${syncPullActiveRefreshSrc}\n${syncPullRebroadcastSrc}\n${syncCutoverSrc}\n${syncDeltaSrc}\n${syncDeltaPlannerSearchSrc}\n${syncDeltaSnapshotSrc}\n${syncDeltaMergeSrc}\n${syncDeltaMergeSearchSrc}\n${syncDeltaRegistrySrc}\n${syncDeltaObservabilitySrc}\n${syncDiagnosticsSrc}\n${syncDiagnoseActionsSrc}\n${syncDiagnoseActionsContextSrc}\n${syncDiagnoseRelayActionsSrc}\n${syncDiagnoseIdentityActionsSrc}\n${syncDiagnoseCutoverActionsSrc}\n${syncDiagnoseUiSrc}\n${syncDiagnoseRenderSrc}\n${syncWindowBindingsSrc}`;
   const exportBlockIncludes = (src, names) => [...src.matchAll(/export\s+\{([^}]*)\};/g)]
     .some(([, block]) => names.every(name => new RegExp(`\\b${name}\\b`).test(block)));
 
@@ -270,13 +275,22 @@ await import('../js/settings.js');
       && syncDeltaMergeSrc.includes('resetPullDeltaSnapshot'));
   assert('service worker precaches sync-delta-merge.js',
     serviceWorkerSrc.includes("'/js/sync-delta-merge.js'"));
-  assert('sync-delta-merge-shapes.js owns pull-side scalar/map/array merge branches',
-    syncDeltaMergeShapesSrc.includes('export async function mergeScalarRowsIntoImported')
-      && syncDeltaMergeShapesSrc.includes('export async function mergeMapRowsIntoImported')
-      && syncDeltaMergeShapesSrc.includes('export async function mergeArrayRowsIntoImported')
-      && syncDeltaMergeShapesSrc.includes('recordPullDeltaSurface'));
-  assert('service worker precaches sync-delta-merge-shapes.js',
-    serviceWorkerSrc.includes("'/js/sync-delta-merge-shapes.js'"));
+  assert('sync-delta-merge-shapes.js owns pull-side merge facade',
+    syncDeltaMergeShapesSrc.includes("from './sync-delta-scalar-merge.js'")
+      && syncDeltaMergeShapesSrc.includes("from './sync-delta-map-merge.js'")
+      && syncDeltaMergeShapesSrc.includes("from './sync-delta-array-merge.js'"));
+  assert('sync delta merge implementation modules own scalar/map/array merge branches',
+    syncDeltaScalarMergeSrc.includes('export async function mergeScalarRowsIntoImported')
+      && syncDeltaMapMergeSrc.includes('export async function mergeMapRowsIntoImported')
+      && syncDeltaArrayMergeSrc.includes('export async function mergeArrayRowsIntoImported')
+      && syncDeltaRowCodecSrc.includes('export async function decodeRowPayload')
+      && syncDeltaMergeSearchSrc.includes('recordPullDeltaSurface'));
+  assert('service worker precaches sync delta merge modules',
+    serviceWorkerSrc.includes("'/js/sync-delta-merge-shapes.js'")
+      && serviceWorkerSrc.includes("'/js/sync-delta-row-codec.js'")
+      && serviceWorkerSrc.includes("'/js/sync-delta-array-merge.js'")
+      && serviceWorkerSrc.includes("'/js/sync-delta-map-merge.js'")
+      && serviceWorkerSrc.includes("'/js/sync-delta-scalar-merge.js'"));
   assert('sync-delta-registry.js owns delta surfaces and identity config',
     syncDeltaRegistrySrc.includes('export const DELTA_ARRAYS')
       && syncDeltaRegistrySrc.includes('export const DELTA_MAPS')
@@ -293,7 +307,7 @@ await import('../js/settings.js');
       && syncDeltaObservabilitySrc.includes('export function resetDeltaTelemetry')
       && syncDeltaObservabilitySrc.includes('export function getDeltaCutoverReadiness')
       && syncDeltaMergeSrc.includes('resetPullDeltaSnapshot(profileId)')
-      && syncDeltaMergeShapesSrc.includes('recordPullDeltaSurface(arrayName'));
+      && syncDeltaMergeSearchSrc.includes('recordPullDeltaSurface(arrayName'));
   assert('service worker precaches sync-delta-observability.js',
     serviceWorkerSrc.includes("'/js/sync-delta-observability.js'"));
   assert('sync-tombstones.js owns remote profile delete helpers',
@@ -882,20 +896,20 @@ await import('../js/settings.js');
   assert('onSyncReceived overlays per-row state AFTER blob merge',
     /merged\s*=\s*localImportedForMerge[\s\S]{0,400}mergeImportedData[\s\S]{0,800}_mergeItemRowsIntoImported/.test(syncPullMergeSrc));
   assert('_mergeItemRowsIntoImported drops tombstoned items from imported arrays',
-    /mergeArrayRowsIntoImported[\s\S]{0,12000}let nextArr\s*=\s*curArr\.filter\(it\s*=>\s*!tombs\.has\(itemIdFn\(it\)\)\)/.test(syncDeltaMergeShapesSrc));
+    /mergeArrayRowsIntoImported[\s\S]{0,12000}let nextArr\s*=\s*curArr\.filter\(it\s*=>\s*!tombs\.has\(itemIdFn\(it\)\)\)/.test(syncDeltaMergeSearchSrc));
   // Resurrection-prevention seed: blob-side `_deleted[arrayName]` must
   // pre-populate the row-side tombs Set, otherwise a peer pushing the
   // row back as live (before pulling our delete) re-inserts it locally.
   assert('_mergeItemRowsIntoImported seeds tombs from local blob _deleted before walking rows',
-    /imported\.\s*_deleted[\s\S]{0,200}\[arrayName\][\s\S]{0,200}tombs\.add/.test(syncDeltaMergeShapesSrc));
+    /imported\.\s*_deleted[\s\S]{0,200}\[arrayName\][\s\S]{0,200}tombs\.add/.test(syncDeltaMergeSearchSrc));
   assert('_mergeItemRowsIntoImported skips inserting items that match a blob-tombstoned itemId',
-    /tombs\.has\(itemId\)\)\s*continue/.test(syncDeltaMergeShapesSrc));
+    /tombs\.has\(itemId\)\)\s*continue/.test(syncDeltaMergeSearchSrc));
   assert('_mergeItemRowsIntoImported prefers per-row payload when itemId already present in array (replace)',
-    /idx\s*!==\s*undefined[\s\S]{0,200}nextArr\[idx\]\s*=\s*item/.test(syncDeltaMergeShapesSrc));
+    /idx\s*!==\s*undefined[\s\S]{0,200}nextArr\[idx\]\s*=\s*item/.test(syncDeltaMergeSearchSrc));
   assert('_mergeItemRowsIntoImported gunzips GZ|v1| payloads via capped variant',
-    /json\.startsWith\('GZ\|v1\|'\)[\s\S]{0,300}_gunzipToStringCapped\(_base64ToBytes\(json\.slice\(6\)\)\)/.test(syncDeltaMergeShapesSrc));
+    /json\.startsWith\('GZ\|v1\|'\)[\s\S]{0,300}_gunzipToStringCapped\(_base64ToBytes\(json\.slice\(6\)\)\)/.test(syncDeltaMergeSearchSrc));
   assert('_mergeItemRowsIntoImported guards against itemId/payload mismatch (defence-in-depth)',
-    /itemIdFn\(item\)\s*===\s*row\.itemId/.test(syncDeltaMergeShapesSrc));
+    /itemIdFn\(item\)\s*===\s*row\.itemId/.test(syncDeltaMergeSearchSrc));
 
   // Snapshot persistence contract
   assert('Delta snapshot key namespaced per (profile, arrayName)',
@@ -1380,9 +1394,9 @@ await import('../js/settings.js');
   assert('_planArrayDelta uses itemIdFn-derived id everywhere (not item.id)',
     /tuples\s*=\s*Array\.isArray\(items\)[\s\S]{0,300}itemIdFn\(it\)/.test(deltaSearchSrc));
   assert('_mergeItemRowsIntoImported uses itemIdFn for replace-or-insert match',
-    /mergeArrayRowsIntoImported[\s\S]{0,8000}DELTA_ARRAY_CONFIG\[arrayName\][\s\S]{0,12000}itemIdFn\(nextArr\[i\]\)/.test(syncDeltaMergeShapesSrc));
+    /mergeArrayRowsIntoImported[\s\S]{0,8000}DELTA_ARRAY_CONFIG\[arrayName\][\s\S]{0,12000}itemIdFn\(nextArr\[i\]\)/.test(syncDeltaMergeSearchSrc));
   assert('_mergeItemRowsIntoImported verifies payload itemId matches row column',
-    /itemIdFn\(item\)\s*===\s*row\.itemId/.test(syncDeltaMergeShapesSrc));
+    /itemIdFn\(item\)\s*===\s*row\.itemId/.test(syncDeltaMergeSearchSrc));
 
   // Live: round-trip the changeHistory itemIdFn — verify a synth itemId
   // for a realistic recordChange entry is allowlist-safe and stable.
@@ -1667,7 +1681,7 @@ await import('../js/settings.js');
   assert('getDeltaCutoverReadiness walks dotted DELTA_ARRAYS entries via getAt',
     /getDeltaCutoverReadiness[\s\S]{0,2000}arrayName\.includes\('\.'\)[\s\S]{0,200}getAt\(importedData,\s*arrayName\)/.test(deltaSearchSrc));
   assert('_mergeItemRowsIntoImported writes nested arrays back via setAt',
-    /mergeArrayRowsIntoImported[\s\S]{0,1600}isNested\s*=\s*arrayName\.includes\('\.'\)[\s\S]{0,400}setAt\(imported,\s*arrayName,/.test(syncDeltaMergeShapesSrc));
+    /mergeArrayRowsIntoImported[\s\S]{0,1600}isNested\s*=\s*arrayName\.includes\('\.'\)[\s\S]{0,400}setAt\(imported,\s*arrayName,/.test(syncDeltaMergeSearchSrc));
   assert('DELTA_SCALARS includes free-form text fields',
     /'interpretiveLens'/.test(deltaSearchSrc) && /'contextNotes'/.test(deltaSearchSrc));
   assert('_planScalarDelta defined',
@@ -1997,8 +2011,8 @@ await import('../js/settings.js');
   assert('_gunzipToStringCapped throws on cap exceeded',
     /total\s*>\s*maxBytes[\s\S]{0,300}refusing to trust/.test(syncPayloadSrc));
   assert('All per-row gunzip sites use the capped variant',
-    /async function decodeRowPayload[\s\S]{0,500}_gunzipToStringCapped\(_base64ToBytes\(json\.slice\(6\)\)\)/.test(syncDeltaMergeShapesSrc)
-      && (syncDeltaMergeShapesSrc.match(/await decodeRowPayload\(row\)/g) || []).length >= 3);
+    /export async function decodeRowPayload[\s\S]{0,500}_gunzipToStringCapped\(_base64ToBytes\(json\.slice\(6\)\)\)/.test(syncDeltaRowCodecSrc)
+      && (syncDeltaMergeSearchSrc.match(/await decodeRowPayload\(row\)/g) || []).length >= 3);
   // Blob path also routes through _gunzipToStringCapped, with the
   // 5 MB MAX_SYNC_PAYLOAD_BYTES cap — a single capped helper is the
   // only gunzip entry point post-2026-05-10 audit (the bare
